feat: 카테고링 기능 구현 및 관심 매물 api 구현 완료

+++ src/main/java/dev/konkuk/home/domain/account/entity/Account.java
@@ -35,7 +35,7 @@ public class Account extends BaseTimeEntity {
             name = "account_properties",
             joinColumns = @JoinColumn(name = "account_id")
     )
-    private List<Integer> properties = new ArrayList<>();
+    private List<Long> properties = new ArrayList<>();
 
     @Builder
     public Account(String name, String email, String imageThumbnail) {
